Added podcast management for admin.
