+++ app/views/admin/podcasts/_form.html.slim
@@ -1,4 +1,4 @@
-= form_for @podcast do |f|
+= form_for [:admin, @podcast] do |f|
   - if @podcast.errors.any?
     #error_explanation
       h2 = "#{pluralize(@podcast.errors.count, "error")} prohibited this podcast from being saved:"
