+++ app/views/admin/podcasts/edit.html.slim
@@ -4,5 +4,4 @@ h1 Editing podcast
 
 = link_to 'Show', @podcast
 '|
-= link_to 'Back', podcasts_path
-
+= link_to 'Back', admin_podcasts_path
