+++ app/views/admin/podcasts/index.html.slim
@@ -19,3 +19,9 @@ table
         td = image_tag podcast.thumbnail.url(:small)
         td = audio_tag(podcast.audio.url, controls: true)
         td = link_to 'Show', podcast
+        td = link_to 'Edit', edit_admin_podcast_path(podcast)
+        td = link_to 'Destroy', [:admin, podcast], data: {:confirm => 'Are you sure?'}, :method => :delete
+
+br
+
+= link_to 'New Podcast', new_admin_podcast_path
