+++ app/views/admin/podcasts/new.html.slim
@@ -1,0 +1,5 @@
+h1 New podcast
+
+== render 'form'
+
+= link_to 'Back', admin_podcasts_path
