--- app/views/podcasts/new.html.slim
@@ -1,5 +1,0 @@
-h1 New podcast
-
-== render 'form'
-
-= link_to 'Back', podcasts_path
