+++ config/routes.rb
@@ -3,10 +3,11 @@ Rails.application.routes.draw do
   root 'podcasts#index'
   devise_for :users
   resources :articles, only: [:index, :show]
-  resources :podcasts
+  resources :podcasts, only: [:index, :show]
 
   namespace :admin do
     root to: 'articles#index'
     resources :articles, except: :show
+    resources :podcasts, except: :show
   end
 end
